refactor #211: TestConfig 중앙화
