+++ src/test/java/com/irum/come2us/global/config/TestConfig.java
@@ -1,0 +1,32 @@
+package com.irum.come2us.global.config;
+
+import com.irum.come2us.domain.auth.application.service.AuthService;
+import com.irum.come2us.domain.deliveryaddress.application.service.DeliveryAddressService;
+import com.irum.come2us.domain.member.application.service.ManagerService;
+import com.irum.come2us.domain.member.application.service.MemberService;
+import org.mockito.Mockito;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
+
+@TestConfiguration
+public class TestConfig {
+    @Bean
+    public MemberService memberService() {
+        return Mockito.mock(MemberService.class);
+    }
+
+    @Bean
+    public ManagerService managerService() {
+        return Mockito.mock(ManagerService.class);
+    }
+
+    @Bean
+    public AuthService authService() {
+        return Mockito.mock(AuthService.class);
+    }
+
+    @Bean
+    public DeliveryAddressService deliveryAddressService() {
+        return Mockito.mock(DeliveryAddressService.class);
+    }
+}
